requirements.txt
streamlit
openai
gspread
google-auth
PyMuPDF
